Refactor to use serializers
